+++ app/src/main/res/layout/activity_main.xml
@@ -28,8 +28,8 @@
 
     <TextView
         android:id="@+id/output_textview"
-        android:layout_width="52dp"
-        android:layout_height="18dp"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
         android:text="TextView"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="@+id/json_change_button"
